moved telemetry sync into telemetry
