--- synchronizer/worker/telemetry/interface.go
@@ -1,6 +1,0 @@
-package telemetry
-
-// SynchronizeEvents interface
-type TelemetryRecorder interface {
-	SynchronizeTelemetry() error
-}
