+++ tasks/telemetrysync.go
@@ -1,19 +1,19 @@
 package tasks
 
 import (
-	"github.com/splitio/go-split-commons/v3/synchronizer/worker/telemetry"
+	"github.com/splitio/go-split-commons/v3/telemetry"
 	"github.com/splitio/go-toolkit/v4/asynctask"
 	"github.com/splitio/go-toolkit/v4/logging"
 )
 
 // NewRecordTelemetryTask creates a new telemtry recording task
 func NewRecordTelemetryTask(
-	recorder telemetry.TelemetryRecorder,
+	recorder telemetry.TelemetrySynchronizer,
 	period int,
 	logger logging.LoggerInterface,
 ) *asynctask.AsyncTask {
 	record := func(logger logging.LoggerInterface) error {
-		return recorder.SynchronizeTelemetry()
+		return recorder.SynchronizeStats()
 	}
 
 	onStop := func(l logging.LoggerInterface) {
